+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v7.0.2 - 2022.08.12 09
+* improve `locale` keys and translations.
+* bump up `@nuogz/i18n` to `1.2.0` and update related code
+
+
 ## v7.0.0 - 2022.08.09 15
 * tweak all files for publishing to npm
 * start use `CHANGLOG.md` since version `v7.0.0`